Cu aptd2y dockerize migration tool (#55)
This pull request adds a Dockerfile for the migration tool. 
I had to move the migration folder out of the monorepo do it nicely (that was seen with @Fenntasy). 
I also edited the circle CI pipeline to push the image to docker hub.

You can see on this screenshot that the image are correctly tagged and pushed to Dockerhub.